services/auth_service/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.33"}
asyncpg = "^0.29.0"
alembic = "^1.13.2"
orjson = "^3.10.7"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
gunicorn = "^23.0.0"
unicorn = "^2.1.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

